feat: Add support headers: artifacts add and task info (#803)
Add the support of headers to `rstuf artifacts add` and `rstuf task
info` command allowing users behind some API authN/Z to send headers
with auth, apikey, or other
requirements

Note: headers are already supported by `rstuf admin` command
