+++ repository_service_tuf/cli/admin/__init__.py
@@ -10,34 +10,12 @@ Provides alternative ceremony, metadata update, and sign admin cli commands.
 """
 from typing import Optional
 
-from repository_service_tuf.cli import click, rstuf
-
-HEADERS_EXAMPLE = (
-    "'Authorization: Bearer <token>, Content-Type: application/json'"
+from repository_service_tuf.cli import (
+    HEADERS_EXAMPLE,
+    _set_settings,
+    click,
+    rstuf,
 )
-
-
-def _set_settings(
-    context: click.Context, api_server: Optional[str], headers: Optional[str]
-):
-    """Set context.obj['settings'] attributes."""
-    settings = context.obj["settings"]
-    if api_server:
-        settings.SERVER = api_server
-    if headers:
-        try:
-            settings.HEADERS = dict(
-                (key.strip(), value.strip())
-                for key, value in (
-                    header.split(":", 1) for header in headers.split(",")
-                )
-            )
-        except ValueError:
-            raise click.ClickException(
-                f"Invalid headers format. Example: {HEADERS_EXAMPLE}"
-            )
-    else:
-        settings.HEADERS = None
 
 
 @rstuf.group()  # type: ignore
